Move all errors to single file
